chore: move LimitDecisions to RateLimiterService
